Beginnings of a real FAT32 driver.
Old driver (FAT12 from u-boot) is now is fat12/ folder for reference purposes. Do NOT use it as we don't have any FAT12 partitions.

Also, beginnings of a README regarding items. Started because the u-boot driver was used since we couldn't tell right off it was a FAT12 driver.

readsw now updates the LED. Use _readsw for non-LED-changing behavior.