Got fat32_open_file further. Need to compare LFNs.
Did some random cleanup of code based on our assumptions.
Just need to create a compare_long_names() that takes a
short directory_entry and a long name and reads the long
name of the directory entry from the disk and compares
the long names. Then, hopefully the open_file will work.