FAT32 driver finally works for read-only access.
I will continue to develop and add write-access. It likely won't be used for the LD, but this is the only direct-access driver available for FAT32 in the open-source world. Also, I plan to use it in other places where I have a memory chip and don't want to write a new FS driver for it.

Modifying drive.c and drive.h to suit a new "drive" would make this driver work with anything else. It will take a little bit of work to make it clean and truly re-usable, but that will all occur in the next commit.